Implement login by Card

+++ src/main/java/com/vietqr/org/dto/TransactionReceiveAdminListDTO.java
@@ -6,7 +6,7 @@ public interface TransactionReceiveAdminListDTO {
 
     String getBankAccount();
 
-    Integer getAmount();
+    Long getAmount();
 
     String getBankId();
